actualizaciones

+++ src/main/java/hn/edu/ujcv/entity/Aeronave/Aeronave.java
@@ -18,7 +18,7 @@ public class Aeronave {
     private String modelo;
     private String fabricante;
     private String aerolinea;
-    private int    tamaño;
+    private double tamaño;
     private int    capacidad;
     private String tipo;
 }
